+++ hipd/close.h
@@ -14,10 +14,10 @@
 #include "lib/tool/pk.h"
 
 int hip_send_close(struct hip_common *msg, int delete_ha_info);
-int hip_handle_close(const uint32_t packet_type,
+int hip_handle_close(const uint8_t packet_type,
                      const uint32_t ha_state,
                      struct hip_packet_context *ctx);
-int hip_handle_close_ack(const uint32_t packet_type,
+int hip_handle_close_ack(const uint8_t packet_type,
                          const uint32_t ha_state,
                          struct hip_packet_context *ctx);
 int hip_purge_closing_ha(hip_ha_t *ha, void *notused);
