+++ hipd/esp_prot_light_update.h
@@ -24,7 +24,7 @@ int esp_prot_send_light_update(hip_ha_t *entry,
                                const int secret_length[],
                                unsigned char *branch_nodes[MAX_NUM_PARALLEL_HCHAINS],
                                const int branch_length[]);
-int esp_prot_handle_light_update(const uint32_t packet_type,
+int esp_prot_handle_light_update(const uint8_t packet_type,
                                  const uint32_t ha_state,
                                  struct hip_packet_context *ctx);
