+++ hipd/output.h
@@ -51,11 +51,11 @@ struct hip_common *hip_create_r1(const struct in6_addr *src_hit,
                                  const struct hip_host_id *host_id_pub,
                                  int cookie_k);
 
-int hip_send_r1(const uint32_t packet_type,
+int hip_send_r1(const uint8_t packet_type,
                 const uint32_t ha_state,
                 struct hip_packet_context *ctx);
 
-int hip_send_r2(const uint32_t packet_type,
+int hip_send_r2(const uint8_t packet_type,
                 const uint32_t ha_state,
                 struct hip_packet_context *packet_ctx);
 
@@ -67,7 +67,7 @@ int hip_send_r2_response(struct hip_common *r2,
 
 int hip_send_i1(hip_hit_t *, hip_hit_t *, hip_ha_t *);
 
-int hip_send_i2(const uint32_t packet_type,
+int hip_send_i2(const uint8_t packet_type,
                 const uint32_t ha_state,
                 struct hip_packet_context *packet_ctx);
