+++ modules/update/hipd/update.h
@@ -66,7 +66,7 @@ int hip_send_locators_to_all_peers(void);
  *
  * @return 0 if succeeded, error number otherwise
  */
-int hip_handle_update(const uint32_t packet_type,
+int hip_handle_update(const uint8_t packet_type,
                       const uint32_t ha_state,
                       struct hip_packet_context *ctx);
